requirements.txt
agno>=1.1.1
yfinance>=0.2.40
python-dotenv>=1.0.0
anthropic>=0.34.0
